befor modify name ParsingFile
